+++ minn-admin.php
@@ -67,6 +67,7 @@ require_once MINN_ADMIN_DIR . 'includes/adapters/simple-301-redirects.php';
 require_once MINN_ADMIN_DIR . 'includes/adapters/eps-301-redirects.php';
 require_once MINN_ADMIN_DIR . 'includes/adapters/query-monitor.php';
 require_once MINN_ADMIN_DIR . 'includes/adapters/scrutoscope.php';
+require_once MINN_ADMIN_DIR . 'includes/adapters/wp-crontrol.php';
 require_once MINN_ADMIN_DIR . 'includes/adapters/cache-purge.php';
 require_once MINN_ADMIN_DIR . 'includes/adapters/updraftplus.php';
 require_once MINN_ADMIN_DIR . 'includes/adapters/disembark.php';